# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cargexo Documentation",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Cargexo API documentation"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth User"
                ],
                "summary": "Login",
                "description": "Login",
                "operationId": "Login",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "admin"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Auth User"
                ],
                "summary": "logout",
                "description": "logout",
                "operationId": "logout",
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "name of the new User",
                        "type": "string",
                        "example": "Ahmed"
                    },
                    "email": {
                        "title": "email",
                        "description": "email of the new User",
                        "type": "string",
                        "example": "email@email.com"
                    },
                    "personal_phone": {
                        "title": "personal_phone",
                        "description": "personal_phone of the new User",
                        "type": "string",
                        "example": "0770123456"
                    },
                    "business_phone": {
                        "title": "business_phone",
                        "description": "business_phone of the new User",
                        "type": "string",
                        "example": "0770123456"
                    },
                    "password": {
                        "title": "password",
                        "description": "password of the new User",
                        "type": "string",
                        "example": "32134"
                    },
                    "type": {
                        "title": "type",
                        "description": "type of the new User",
                        "type": "string",
                        "example": "admin"
                    },
                    "address": {
                        "title": "address",
                        "description": "address of the new User",
                        "type": "string",
                        "example": "Baghdad/alkaradaa 62 Street"
                    },
                    "date": {
                        "title": "date",
                        "description": " signIn date or hiring date ",
                        "type": "string",
                        "example": "2/2/2020"
                    },
                    "status": {
                        "title": "status",
                        "description": "status of the new User",
                        "type": "string",
                        "example": "active"
                    },
                    "is_deleted": {
                        "title": "is_deleted",
                        "description": "is_deleted determain if the user deleted or not",
                        "type": "boolean",
                        "example": "true"
                    },
                    "note": {
                        "title": "note",
                        "description": "note about the user",
                        "type": "boolean",
                        "example": "true"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            }
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}